Move SimpleTracingPolicy class back to program
