+++ samples/CloudClipboard/GarbageCollector/Program.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Azure.Core;
+using Azure.Core.Pipeline;
 using Azure.Identity;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
@@ -26,6 +27,23 @@ namespace GarbageCollector
                 BlobContainerClient containerClient = serviceClient.GetBlobContainerClient(container.Name);
                 await containerClient.DeleteAsync();
             }
+        }
+    }
+
+    public class SimpleTracingPolicy : HttpPipelinePolicy
+    {
+        public override async ValueTask ProcessAsync(HttpMessage message, ReadOnlyMemory<HttpPipelinePolicy> pipeline)
+        {
+            Console.WriteLine($">> Request: {message.Request.Method} {message.Request.Uri}");
+            await ProcessNextAsync(message, pipeline);
+            Console.WriteLine($">> Response: {message.Response.Status} from {message.Request.Method} {message.Request.Uri}\n");
+        }
+
+        public override void Process(HttpMessage message, ReadOnlyMemory<HttpPipelinePolicy> pipeline)
+        {
+            Console.WriteLine($">> Request: {message.Request.Uri}");
+            ProcessNext(message, pipeline);
+            Console.WriteLine($">> Response: {message.Response.Status} {message.Request.Method} {message.Request.Uri}");
         }
     }
 }
